Added some reset functions and called when loading file
Before the undo/redo stacks wouldn't be cleared when loading a new file. Now functions were added in main and undo/redo handler so that a single function can be called to reset main to its original state. This is called when loading data from a file. and can be called later with a 'Reset All' menu item or something.

story:95