Deeper undo/redo test for special person deletion with fix
More reference problems arose with the new test. Spent a while fixing them. The fix was to make the Team cloning function make a nested clone with the members, but that in turn made a stack overflow since the Person and Team clone functions call each other. So I made a new clone function for each of them to avoid this.

story:113