Unit testing for reverting
had to add a function in main to set the primary stage so that the tests wouldnt have a null pointer exception because the app wasnt actaully running.
added unit tests for single case, empty case and multiple case. all loading,saving and save as test cases will be done manually as it would take alot of time to set it up so they could be unit tested and it wouldnt be quicker to just manual test them.

Story:108