Clean up project controller and skill removal case for person dialog with bugfix
Project controller had two listeners for the same field. Moved the content of the second listener to the first. Removed unused functions. Bug fix for removing skills from people since it did not refresh the save button. Also added an extra case for removing skills from people when they are assigned to a role which relies on that skill.

story:109