made creating stories with ACs work.
LMPC changed to display the ACs
ACDialogController changed so that it will return out the ac into the SDC
StoryDialogController changed to have a list to hold the acs and put them into the listview. Creating now works with this. EDITING TO YET TO IMPLEMENT.
Removed useless function from Story
The ACDialog fxml changed to wrap text and meet design trends.

story:101

pair[adl52]