Added Estiamtes to Backlog.
Added an estimates variable to the backlog class to allow it to hold one, added getter and setter for it too (for displaying and seting durr)

Changed BacklogDialogController so now the estimates combobox has all the possible scales to choose from and you must choose and save a backlog with the scale (As AC says)

Did the greying out of the save button on editing for the estimation combobox.

LMPC changed so the chosen esitmation scale is displayed when the backlog is selected. This included a work around for loading old files to just display "null" as they wont have an estimation scale selected.

added null to alot of test cases so that this could compile as you must have an estimation scale so I changed the constructor to take one in. I put TODO next to the ones I changed in case its needed to be adapted later.

Changed the dialog's estimation combobox to have a name in line with our others

story:100