Found a bug with loading with estimates not existing. Fixed it.
Shingy found a bug with loading things. Figured out it was a problem with estimates not being loaded, since I forgot where the current main was cleared on loading. We decided to actually load the estimates since we load roles anyway. Doing that next.

Before that started on beefing up the estimate java docing

pair[ssc55]
story:100