Implement cascading delete for backlogs.
Updated the undo/redo handler for backlogs so that stories are returned after their associated backlog is deleted and undone/redone. Updated main, to show an alert that warns the user that a cascading delete will be performed if they continue.
Also updated the affected funtions javadocs

Story:252
pair[cab183]