Created sync function for Projects and backlogs. Fixed toolbarquickadd button image to be perfectly balanced
Created the sync function so that projects and backlogs are referencing properly after a revert. This function is yet to be used anywhere as i cannot test it until earlier tasks are completed but i will test it once those are done and use it.
Included the Javadoc for the sync function to explain its use.
Fixed the image used for the immediate toolbar so that it is balanced and more aesthetic.

Story:252,253