Added a visual representation for estimates in the story dialog
I edited the story dialog and its controller so that the estimate that the story has is shown in its dialog. If it has no estimate or is currently being created then it simply says "No Estimate". It has a tooltip to indicate it cannot be edited here and should be done in the backlog. This can be changed later if wanted.

Story:99