Test project with a backlog reverting
I wrote some unit tests for this. Took a while extra as i found some bugs and then all our unit tests broke for a while. Also had a problem figuring out a strange happening with revert in the unit tests where it doesnt change the referenced object after reverting but this is handled in the application itself.

Story:251