Sprint Model Unit testing
Wrote up some unit tests for the new Sprint model.
These tests include setting up a Sprint object, testing the toString method and the equals method.
Changed the doc slightly in the sprint model to show that the sprint goal needs to be unique in the main list of all sprints.

Story:235