Sprint creation/edit dialog
I designed and created the dialog for creating or editing a sprint.
The only field that is not allowed to be null is the sprint Goal.
I have used the term "Sprint Goal" everywhere that "Sprint label" usually would. This is because in the acceptance criteria it specifys it as a sprint goal which works as a label. This is subject to change tho.

Story:235