Fixed a readiness undo/redo bug.
If a story is marked as ready, then removed from its assigned backlog or marked as not-estimated, it becomes not ready. If you then undo the backlog removal/estimate removal, the story was still marked as not ready instead of being returned to its ready state.

Added a list to backlog controller that stores which stories will be affected which is then passes to the undo/redo handler and added to the undo/redo stack as an agile item. May have to be modified in the future if similar situations come up, but for now it works.

Story: 251