Checking for cyclic dependency implementation & test.
Changed the implementation of the cyclic dependency checker as it was not working in some cases so some extra logic had to be added.
Unit tests have now been written for them testing for multiple cases. All of them pass of course.
All Jdocs have been added to functions and to ambiguous tests that need explaining on their use. The Jdoc in the tests shows a graph of the set up. Its pretty cool.

pair[ssc55]

Story:98