Redesigned the UI for dependency dialog
The dialog before was decided to be unintuitive and needed to be redesigned so we did. Added an extra list for selecting the subject story and save/cancel buttons for consistency. This included editing functions in the controller to make it useable with the new list.
Javadocs were updated on the refreshList and initialize list funtions. others aswell.

pair[msr51]

Story:98