Project model function to check current teams. Sprint dialog somewhat properly populates team/release combo boxes
New function in project model to get teams currently allocated based on start/end dates. The sprint dialog was changed to react to changes in the backlog field. Initially the team and release combo boxes are disabled and are only enabled once a backlog which is allocated to a project is selected. This uses current fields from that project to populate those combo boxes.

story: 235