Adding and removing stories from sprints in the dialog.
Implemented functionality to add or remove stories to and from a sprint while maintaining priority order using an interesting fix. The fix involves maintaining both a set and a prioritsed list for allocated stories. On refreshing, the available stories resets itself with stories from the backlog in the original order and then removes all stories which exist in the set. The allocated stories list refreshes by resetting it self with stories from the backlog in the original order and then removes all stories which do not exist in the set.

story: 235
pair[msr51]