Validation for sprint dialog input fields
Checks that sprint goal is not empty and is unique. Check that start date is not null and before release date. Check that end date is not null, after start date and before release date. Checks that all combo values are not null and the project is not null. Everything can be empty.

story:235
pair[apw76]