Added undo/redo capabilities in the application for sprints
Changed the Sprint controller to allow for undoing and redoing. This included making a new function and editing the confirm button.
Changed the main class so that Sprints could be deleted as someone forgot to add to the switch statement.
Also slightly edited a function in main so that there was no redundant assigning to a release in the showReportDialog function.
Javadoc was update where needed and added for the new function.

Story:235