Changed which story is selected after moving a story between the two Sprint story lists.
Now, instead of selecting the first story in the list once a story has been moved from the available stories list to the allocated stories list, or from the allocated stories list to the available stories list, the story that is next in the list is selected. This means that you can quickly move stories around if they are already in the correct order in the list.

story: 235