Added check for story readiness.
A story now cannot be assigned to a sprint if it is not ready.
Added proper behaviour when a story is marked as not ready from a nested dialog inside a sprint.
A story can not be marked as not ready if it is assigned to a sprint.
A story is removed from available sprint stories if it is marked as not ready.

As a side effect, story details now properly update when changed from a nested dialog inside of a sprint.

Story: 235