Cascading delete for projects and sprint
When you delete a project with a sprint assigned to it, the error dialog will also mention any assigned sprints that will also be deleted. These sprints are deleted along with the releases and undo/redo has been refactored to handle this case as well. NB: wish we learned about the composite design pattern and open closed principle earlier.

story:235