Updated story model for greying out of button. Updated story diplay pane
Updated the story model so that the button greys out when the status remains unchanged along with the old requirements for this to be greyed out.
Updated the story display pane so that it will display what status the story is in in plain text for the user.
Model an extra getter to get the status as a sting rather than an ENUM object. Its kinda bad tho should be looked at later maybe.

Story:325