Added extra values to estimation scales.
Added "Epic Dino" to Dinos scale.
Added "21" to Fibonacci scale.

Currently, the old scales are used upon loading a file. This can be changed if required.

A JavaFX warning?/error? occured a single time while testing this. I was not able to repeat it, however it was something along the lines of 'index exceeds maxCellCount'.

Story: 317