Can create/remove tasks from story dialog
Ended up removing the Taskable interface from the task dialog controller but it will still be used in undo/redo. The new implementation allows for more temporary changes for smoothly cancelling transactions. Removed the limitation of story of being in a sprint before having tasks which would've happened if I carried on my old way.

story:314