Created barebones Toolbar FXML.
Created the Toolbar.fxml file that contains the very basic stuff needed for the toolbar. Everything is placeholder at the moment. Text will be replaced with images and more buttons will be added.

Toolbar.fxml will be added to the Main Pane through the Main class in the same way that the Menu Bar is.

Story: 312