Updated the thumbnail, added tasks to reports, wrote unit tests for reports
I updated the dino thumbnail so there is no random white on the left of it so it looks nicer. Wrote tasks into reports so it will be displayed inside stories and sprints now when reporting. Quite sure this works will double check once all the implementation is done. Wrote the unit tests for this to ensure that it is still writing the report file.

Story:314