List cell in Task dialog includes text field for logging effort later
Introduced text field in allocated people list. Spent a while because I was copying code from the wrong commit in the past based on the Edit button on AC list implementation. I was copying from the one without the proper alignment of the button. It's not linked to anything yet.

story:314