Fixed undo/redo for scrum board and unit tests
Unit tests broke because scrumboard controller functions were being called from within the handler which is null during tests. The actual undo from the scrum board dialog was broken because the undo redo object was global and not reinitialised with each new transaction. Note this was Su-Shing and Liang just working on Craig's laptop.

#pair: zma16, ssc55
#story:315