Fixed bugs so that revert works with scrum board and fixed add button
The issue was that the temporary nonStory in the controller had multiple instances which made it hard to detect if it was selected. For some reason after reverting before the sprints would double for each backlog as well within the scrum board combo box. All kinds of weird behaviour was occurring. The solution to these problems was to only ever have one instance of nonStory even after reverting. This meant removing all local nonStorys and all initalisers apart from the first. In doing so, we also fixed the bug with the new task button when sprint is selected.

Story: 315
pair[ssc55]