Fixed undo/redo for scrum board task edit and image on new task button
UndoRedo object is now put onto the stack after editing a task by double clicking on it in the scrum board. And threw in a button for the new task button on the scrum board and put the scrum board into an anchor pane (but not very effective at the moment). Fixed bug where scrum lists weren't refreshing when closing sprint/story dialogs.

Story: 315
pair[ssc55]