Added impediments to tasks
Added impediments to the task model and the task dialog. Added functionality to populate task dialog on edit, save on create or edit, and disable confirm button on no changes in dialog. Made sure it worked for revert, save/load, reports, undo/redo and made it work if it didn't.

story:314
pair[zma16]