Created Hyperlink generator function
Added a function which takes in an AgileItem and generates a Hyperlink (or rather pseudo-hyperlink with a Text object) to change the currently displaying list type and select the target AgileItem. Pseudo-hyperlinks were used because Hyperlinks had different line spacing to Text. Added several helper functions to make this work nicer.

story:365