Added accordion for task dialog to display a list of logging effort.
Discussed with other team members on how to display the list of logging effort. The outcome is to have a accordion in place and sort the effort spent by person, and drop down a list of dates and times has been logged against that person, and further drop down to the comments made for that particular logging effort.

Story: 357