Added context menu to the listview so that when right click on an item in the list, we can either edit that item or create a new item.
Discussed with other team members on how to implement the menu items, added the context menu to the list, either create new or edit current item. However, after the implementation we realised that the edit function is kind of redundant because of the double-click-edit, and we can't really create new if click on the empty space within the list. So we have to scrap this idea and decided to just use the plus button that we already had for item creation.

Story: 361