Basic implementation of creating effort and updated model
Implemented the effort dialog controller to be able to create effort. It populates with current date and time on first opening and contains validation when confirming effort creation.  Updated task dialog controller so that effort can be added to a "task" on creation. Updated effort dialogs and effort model to better suit desired functionality.

story:357 pair[adl52]