Custom columns for effort table in task dialog
Customised the date and time column to only show the date/time in the format "dd/MM/yyyy H:mm" rather than "yyyy-MM-ddTH:mm" or whatever it was. Also customised the spent effort column to show the formatted duration string rather than the number of minutes in integer form.

story:357