Changes to efforts are undone on cancelling task dialog
When cancelling the task dialog after making changes to the efforts, the changes are now undone when closing the task dialog with a confirmation dialog. However I found that the efforts are completely cleared when cancelling the task edit from the parent sprint dialog... I haven't made any progress on that problem yet.

story:357