Fixed bug with Task cloning constructor
The cloning constructor set the reference  of the efforts list to be the one of the effort that is to be cloned rather than just copying its values. This fixed a bug so now when cancelling the sprint dialog after making changes to a task by making a change to its efforts, the  changes to the efforts are now undone when closing the sprint dialog.

story:357