Creating and editing product owners from backlog dialog
Implemented a mode in the person dialog to handle differently when opened from the backlog dialog. This makes the product owner skill automatically be allocated to the person and cannot be removed to ensure the person appears in the available product owners combo box. This dialog mode is accessed from the backlog dialog via the create and edit buttons beside the combo box.

story:357
pair[zma16]