Added checks for nested dialogs so that dialogs cannot be opened in a circular fashion.
Added an ArrayList openControllers that stores all currently open controllers. As a new controller is opened, it is added to the list and as a controller closes it is removed from the list. When a new dialog is opened, it is checked against the openControllers list and the dialog won't open if it exists in the list so that duplicate windows cannot be opened.

Also created a new interface AgileController that allows a controller to be passed around generically.

story: 361