Fixed a bug with story estimates in main pane
The observable list from the story dialogs were directly referenced in the story objects. This fired the listeners in the story dialog controller. The fix was to make this no longer observable and made the setter to be a pseudo-setter by clearing the stored list and copying the new values.

story: 361
pair[ssc55]